types: Move Evaluable from FD to types
