+++ devito/operator/operator.py
@@ -9,7 +9,6 @@ import ctypes
 from devito.archinfo import platform_registry
 from devito.compiler import compiler_registry
 from devito.exceptions import InvalidOperator
-from devito.finite_differences import Evaluable
 from devito.logger import info, perf, warning, is_log_enabled_for
 from devito.ir.equations import LoweredEq
 from devito.ir.clusters import ClusterGroup, clusterize
@@ -25,7 +24,7 @@ from devito.passes import Graph, NThreads, NThreadsNested, NThreadsNonaffine
 from devito.symbolics import estimate_cost
 from devito.tools import (DAG, Signer, ReducerMap, as_tuple, flatten, filter_ordered,
                           filter_sorted, split, timed_pass, timed_region)
-from devito.types import CustomDimension
+from devito.types import CustomDimension, Evaluable
 
 __all__ = ['Operator']
