+++ devito/types/__init__.py
@@ -1,6 +1,7 @@
 # Need only from within Devito
 from .basic import *  # noqa
 from .array import *  # noqa
+from .lazy import *  # noqa
 
 # Needed both within and outside Devito
 from .dimension import *  # noqa
