+++ devito/types/lazy.py
